refactor(modules): Remove getModuleContent and getModuleContentWithoutTransforms
Both of these methods relied on the formerly impure getSchemaContent, which was mutating state invisibly in the pipeline. This removes those functions and exports the pure functions for use in the actual business logic where state is modified. This will require changes to all of the SDK's utility scripts, but such is life. Those changes forthcoming.