Add static Extensions module
This module exposes the APIs used to do Extension SDK registration and initialization so that Apps can pass them configuration without worrying about which order the SDKs load and Extension SDKs can get access to that config, as well as platform tokens.

There is also a hard coded approvePurchase dependency, for the Commerce SDK. We'll need to remove this once we do user grants.

it's wrapped in future-proofed code that needs testing.