fix: Incorrect reference to non-existing type in firebolt.d.ts
The fix specifically addresses the `SecondScreenEvent` type within the
`SecondScreen` module. When the module is generated, `schema.title` is
set, causing `theTitle` (with a value of 'SecondScreenEventType') to be
interpreted as a type. This prevents the expansion of an enum defined in
the schema, unlike what occurs on the `next` branch.